Seem to have a simple but solid and working modal confirmation for End ladder process. Will probably add it to Start Ladder but also to Remove Team and also Delete League and think of other things for it but this is really all I wanted modals for. Will look into other interactive components later too.
